refactor(test): move fixture to conftest.py Also: - Remove redundant get_user method from CasedbTestClient - Introduce get_user method in test classes for consistent user retrieval - Update test cases to utilize the new get_user
